+++ app/src/main/java/com/ioannapergamali/mysmartroute/view/ui/components/NavigationDrawer.kt
@@ -7,22 +7,29 @@ import androidx.compose.material.icons.filled.ExitToApp
 import androidx.compose.material.icons.filled.Logout
 import androidx.compose.material.icons.filled.Storage
 import androidx.compose.material.icons.filled.AdminPanelSettings
-import com.ioannapergamali.mysmartroute.view.ui.components.LogoImage
-import com.ioannapergamali.mysmartroute.view.ui.components.LogoAssets
+import androidx.compose.material.icons.filled.Person
+import androidx.compose.material.icons.filled.Login
 import com.google.firebase.auth.FirebaseAuth
 import androidx.compose.material3.*
 import androidx.compose.material3.MaterialTheme
 import androidx.compose.ui.text.font.FontWeight
 import androidx.compose.foundation.layout.Row
+import androidx.compose.foundation.layout.Spacer
+import androidx.compose.foundation.layout.size
 import androidx.compose.foundation.layout.padding
 import androidx.compose.runtime.Composable
+import androidx.compose.runtime.mutableStateOf
+import androidx.compose.runtime.remember
+import androidx.compose.runtime.DisposableEffect
 import androidx.compose.runtime.rememberCoroutineScope
 import androidx.compose.ui.Modifier
+import androidx.compose.ui.Alignment
 import androidx.compose.ui.unit.dp
 import androidx.navigation.NavController
 import androidx.compose.ui.platform.LocalContext
 import android.app.Activity
 import kotlinx.coroutines.launch
+import com.google.firebase.firestore.FirebaseFirestore
 
 @Composable
 fun DrawerMenu(navController: NavController, closeDrawer: () -> Unit) {
@@ -31,25 +38,46 @@ fun DrawerMenu(navController: NavController, closeDrawer: () -> Unit) {
         drawerContentColor = MaterialTheme.colorScheme.onSurface,
         drawerTonalElevation = 4.dp
     ) {
-        Row(modifier = Modifier.padding(16.dp)) {
-            LogoImage(
-                drawableRes = LogoAssets.LOGO,
-                contentDescription = "logo",
-                modifier = Modifier.padding(end = 8.dp),
-                size = 36.dp
+        val user = FirebaseAuth.getInstance().currentUser
+        val username = remember { mutableStateOf<String?>(null) }
+
+        DisposableEffect(Unit) {
+            val auth = FirebaseAuth.getInstance()
+            val listener = FirebaseAuth.AuthStateListener { firebaseAuth ->
+                val uid = firebaseAuth.currentUser?.uid
+                if (uid != null) {
+                    FirebaseFirestore.getInstance().collection("users")
+                        .document(uid)
+                        .get()
+                        .addOnSuccessListener { doc ->
+                            username.value = doc.getString("username")
+                        }
+                } else {
+                    username.value = null
+                }
+            }
+            auth.addAuthStateListener(listener)
+            onDispose { auth.removeAuthStateListener(listener) }
+        }
+
+        Row(
+            modifier = Modifier.padding(16.dp),
+            verticalAlignment = Alignment.CenterVertically
+        ) {
+            Icon(
+                imageVector = Icons.Filled.Person,
+                contentDescription = null,
+                modifier = Modifier.size(36.dp),
+                tint = MaterialTheme.colorScheme.primary
             )
+            Spacer(Modifier.size(8.dp))
             Text(
-                "Menu",
+                text = username.value ?: if (user != null) user.email ?: "" else "Guest",
                 style = MaterialTheme.typography.titleMedium.copy(fontWeight = FontWeight.Bold)
             )
         }
-        val context = LocalContext.current
-        val user = FirebaseAuth.getInstance().currentUser
 
-        user?.email?.let { email ->
-            Text(email, modifier = Modifier.padding(start = 16.dp, bottom = 8.dp))
-            Divider(color = MaterialTheme.colorScheme.outline)
-        } ?: Divider(color = MaterialTheme.colorScheme.outline)
+        Divider(color = MaterialTheme.colorScheme.outline)
 
         
         NavigationDrawerItem(
@@ -100,6 +128,18 @@ fun DrawerMenu(navController: NavController, closeDrawer: () -> Unit) {
                     closeDrawer()
                 },
                 icon = { Icon(Icons.Filled.Logout, contentDescription = null, tint = MaterialTheme.colorScheme.primary) }
+            )
+        } else {
+            NavigationDrawerItem(
+                label = { Text("Login") },
+                selected = false,
+                onClick = {
+                    navController.navigate("home") {
+                        popUpTo("home") { inclusive = true }
+                    }
+                    closeDrawer()
+                },
+                icon = { Icon(Icons.Filled.Login, contentDescription = null, tint = MaterialTheme.colorScheme.primary) }
             )
         }
         val activity = (LocalContext.current as? Activity)
